Update phasmcore from monorepo (1000 commits since 56a8738)
## Changes since 56a8738:

- core/cli: bump version 0.2.4 → 0.2.5 (cargo-dist workspace parity)
- core: bump version 0.2.4 → 0.2.5
- core/tests: Task #209 — add f60 PSNR regression gate
- core: provenance + license audit follow-up — clean-room hygiene refresh
- core/h264: stego — default to BRdoConfig::PRODUCTION_VISUAL in build_encoder
- core/h264: Task #208 — fix B-frame residual stego hook missing in emit_b_residual_for_pred
- core/h264: Task #204 Phase 1 — typed BRdoConfig on Encoder + b_rdo_enabled_with(config)
- core/h264: Task #207 — fix walker neighbor commit divergence for B_8x8 / Partitioned
- core/h264: §B-Partitioned-Residual Stage D (#206) — walker un-reject + force-MV plumbing
- core/h264: §B-Partitioned-Residual Stages B+C (#206) — wire residual into emit_b_partitioned + emit_b_8x8
- core/h264: §B-Partitioned-Residual Stage A (#206) — extract emit_b_residual_for_pred
- core/h264: bisect harness + locate B-frame 4x4 speckle root cause
- Revert "core/h264: WET-mark B-slice MVD positions to eliminate B-frame stego artifacts"
- core/h264: WET-mark B-slice MVD positions to eliminate B-frame stego artifacts
- core/h264: enable B_RDO + B_RESIDUAL on iPhone7 demo (closes #195, partial #203)
- core/h264: B-direct-streaks (#196) FIXED -- implement spec 8.4.1.2.2 step 6 colZeroFlag
- core/tests: switch iPhone7 demo to IPPPP pending #196 (B-direct streak ghosting)
- core/tests: mux iPhone7 demo via phasm build_mp4_with_pattern (fix stutter)
- core/h264: revert iPhone7 demo to no-RDO pending #195 (B-recon divergence)
- core/h264: section B-direct-fix.v2 (#194) FIXED — emit transform_size_8x8_flag in B-frame inter residual emission
- core/h264: section B-direct-fix.v2 (#194) — fix PHASM_B_FORCE_MODE bypass + tighten ffmpeg gate
- core/h264: cleanup stale section 6E-A6/6E-A6.1q TODO comments (#190)
- core/h264: §B-direct-fix.v2 (#194) — disable B-RDO in iPhone7 demo + extend PSNR regression to 1080p
- core/h264: fix B-frame slice header frame_num — use preceding-ref value (spec § 7.4.3)
- core/h264: pair B-residual emission with B-RDO + add visual PSNR regression test (#187)
- core/tests: restore h264_iphone_demo.rs PHASM_B_RDO=1 (was changed during bisect)
- core/h264: §Default-flip.IBPBP (#193) — production default IPPPP → IBPBP
- core/h264: §6E-D.5(o) — Bi rate hack (5→2 bins) + 0.9× variance-reduction discount
- core/h264: §6E-D.5(m) — HF-proportional Direct distortion multiplier
- core/h264: §B-direct-fix root cause — drop bucket-derived L0/L1/Bi from no-RDO fallback
- core/h264: §B-direct-fix stage 1.6 — deterministic-seed env var for diag
- core/h264: §B-direct-fix stage 1.5 — fast-iterate PSNR diag harness
- core/h264: §B-direct-fix stage 1 — ColocatedMvGrid foundation
- core/h264: HOTFIX — flip default GopPattern IBPBP→IPPPP (B-frame visual bug)
- Revert "core/h264: §6E-D.5(m) — HF-proportional Direct distortion multiplier"
- Revert "core/h264: §6E-D.5(o) — Bi rate hack (5→2 bins) + 0.9× variance-reduction discount"
- core/h264: add iPhone7 demo encode test (visual inspection harness)
- core/h264: §6E-D.5(o) — Bi rate hack (5→2 bins) + 0.9× variance-reduction discount
- core/h264: §6E-D.5(m) — HF-proportional Direct distortion multiplier
- Website: H.264 post ship-state corrections + AlternativeTo in Organization sameAs
- core/h264: §Stealth.L4.6.5 — close cluster: PPS now byte-equivalent to x264-medium
- core/h264: §Stealth.L4.6.4 — weighted_pred_flag = 1 + degenerate pred_weight_table
- core/h264: §Stealth.L4.6.3 — chroma_qp_index_offset + second_chroma_qp_index_offset = -2
- core/h264: §Stealth.L4.6.2 — num_ref_idx_l0_default_active_minus1 = 2 + slice override
- core/h264: §Stealth.L4.6.1 — pic_init_qp_minus26 0 to -3 (init QP 26 to 23)
- core/h264: §6E-D.5(l) — PPS weighted_bipred_idc 0→2 (container stealth match)
- docs: mark iOS App Clip Experience as configured
- docs: IMPLEMENTATION_PLAN.md — mark items as done/partial after audit
- core/h264: §6E-D.5(k) — Direct multiplier ×2.5 + drop Bi early-termination
- core/h264: §6E-D.5(j) — Direct distortion multiplier 2 → 3
- core/h264: §6E-D.5(i) — Direct distortion multiplier (no-residual penalty)
- core/h264: §6E-D.5(f)+(g) — calibration-style closing of Direct gap
- Website: tint Latest-section cards with the same green-rgba bg as the pillar so the cards read as cards
- Website: blog magazine-layout redesign + 4 category landing pages
- core/h264: §6E-D.5(e) — drop artificial skip_cost from RDO; Skip only via CBP-zero pre-check
- Website: deploy Photoshop-resized H.264 CABAC hero
- Website: publish H.264 CABAC cascade post
- core/h264 + scripts: §6E-D.5(d) diagnostics + #170 histogram script enhancement
- core/h264: §6E-D.6(c) follow-up — update header comment + REFERENCE_DISTRIBUTIONS index
- core/h264: §6E-D.6(c) — re-baseline gate reference to same-fixture x264
- core/h264: §6E-D.5(c) — bipred-aware joint ME refinement (subme≥7 style)
- docs/h264: §6E-D doc update — trace-breakthrough analysis + revised plan
- Website: deploy Photoshop-resized H.264 hero
- Website: rebuild H.264 post with 3 inline diagrams embedded
- Website: publish H.264 video stego post — encoder + container fingerprint mimicry
- core/h264: §6E-D.5(d) — Bi early-termination threshold
- core/h264: §6E-D.5b Tracks A+B — Skip pre-check + CABAC-aware rate weighting
- core/h264: §6E-D.5(a) — spec-correct spatial-direct list usage in RDO
- core/h264: §6E-D.5 — wire fast-RDO B-mode decision behind PHASM_B_RDO
- core/h264: §6E-D.3 — B_8x8 (mb_type 22) candidates in fast RDO
- core/h264: §6E-D.2 — Partitioned 16x8 + 8x16 candidates in fast RDO
- core/h264: §6E-D.1 — switch to fast RDO (SATD-domain) as default
- core/h264: §6E-D.1 — B-slice RDO foundation (16x16 family)
- core/h264: re-frame §6E-A6.5 partition-shape gap → blocked on Phase 6E-D RDO
- core/h264: §6E-A6.5 round 3 — direction probe + v3 calibration regress + final lock
- core/h264: §6E-A6.5 calibration round 2 — 1080p probe + 3-variant comparison
- core/h264: §6E-A6.5 calibration probe + comment update (#155, #156 prep)
- core/h264: §6E-A6.5 (#155) — stealth distribution gate
- core: clippy cleanup — 58 warnings → 0
- core/h264: §6E-A6.1q.e (#154) — per-list within-MB MVD ctxIdxInc

... (920 more monorepo commits omitted; see monorepo log for full range)

Range: 56a8738..9b07326 (2026-05-04)

[squash of 1000 monorepo commits]

+++ src/stego/mod.rs
@@ -193,7 +193,7 @@ fn smart_decode_inner(stego_bytes: &[u8], passphrase: &str) -> Result<(PayloadDa
     // bar jumping back to 0%.  The bar continues smoothly from the Armor
     // phase into the Ghost phase.
     let (armor_done, _) = progress::get();
-    progress::set_total(armor_done + GHOST_DECODE_STEPS as u32);
+    progress::set_total(armor_done + GHOST_DECODE_STEPS);
     let ghost_result = ghost_decode(stego_bytes, passphrase);
     match ghost_result {
         Ok(payload) => return Ok((payload, DecodeQuality::ghost())),
